backend/requirements.txt
asgiref==3.2.10
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
requests==2.26.0
django==2.2.19
djangorestframework==3.12.4
drf-extra-fields==3.4.0
gunicorn==20.0.4
idna==3.3
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
pycparser==2.21
psycopg2-binary==2.8.6
PyJWT==2.1.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework-simplejwt==4.8.0
django-filter==2.4.0
pytz-2022.7.1
sqlparse-0.4.3
Pillow-9.4.0
djoser==2.1.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.11
zipp==3.8.1
python-dotenv==0.21.1